+++ src/field/focused_laser.rs
@@ -8,7 +8,7 @@ use crate::geometry::{FourVector, StokesVector};
 use crate::nonlinear_compton;
 use crate::pair_creation;
 
-use super::{RadiationMode, EquationOfMotion};
+use super::{RadiationMode, EquationOfMotion, Envelope};
 
 /// Represents the envelope of a focusing laser pulse, i.e.
 /// the field after cycle averaging
@@ -19,10 +19,12 @@ pub struct FocusedLaser {
     wavevector: FourVector,
     pol: Polarization,
     bandwidth: f64,
+    envelope: Envelope,
 }
 
 impl FocusedLaser {
-    pub fn new(a0: f64, wavelength: f64, waist: f64, duration: f64, pol: Polarization) -> Self {
+    pub fn new(a0: f64, wavelength: f64, waist: f64, n_cycles: f64, pol: Polarization) -> Self {
+        let duration = n_cycles * wavelength / SPEED_OF_LIGHT;
         let wavevector = (2.0 * consts::PI / wavelength) * FourVector::new(1.0, 0.0, 0.0, 1.0);
         FocusedLaser {
             a0,
@@ -31,13 +33,35 @@ impl FocusedLaser {
             wavevector,
             pol,
             bandwidth: 0.0,
+            envelope: Envelope::Gaussian,
         }
     }
 
-    pub fn with_finite_bandwidth(self) -> Self {
+    pub fn with_envelope(self, envelope: Envelope) -> Self {
         let mut cpy = self;
-        let n_fwhm = SPEED_OF_LIGHT * cpy.duration * cpy.wavevector[0] / (2.0 * consts::PI);
-        cpy.bandwidth = (0.5 * consts::LN_2).sqrt() / (consts::PI * n_fwhm);
+        cpy.envelope = envelope;
+        cpy
+    }
+
+    /// Returns the number of wavelengths corresponding to the pulse
+    /// duration
+    #[inline]
+    fn n_cycles(&self) -> f64 {
+        SPEED_OF_LIGHT * self.duration * self.wavevector[0] / (2.0 * consts::PI)
+    }
+
+    pub fn with_finite_bandwidth(self, on: bool) -> Self {
+        let mut cpy = self;
+        let n_fwhm = match cpy.envelope {
+            // n_fwhm = 2 n acos[1/2^(1/4)] / pi
+            Envelope::CosSquared => 0.36405666377387671305 * cpy.n_cycles(),
+            Envelope::Flattop | Envelope::Gaussian => cpy.n_cycles(),
+        };
+        cpy.bandwidth = if on {
+            (0.5 * consts::LN_2).sqrt() / (consts::PI * n_fwhm)
+        } else {
+            0.0
+        };
         cpy
     }
 
@@ -47,6 +71,38 @@ impl FocusedLaser {
 
     fn rayleigh_range(&self) -> f64 {
         0.5 * self.wavevector[0] * self.waist.powi(2)
+    }
+
+    /// Returns the mean-squared pulse envelope ⟨f^2(ϕ)⟩ and its gradient
+    /// d⟨f^2(ϕ)⟩/dz at the given phase ϕ
+    fn envelope_and_grad(&self, phase: f64) -> (f64, f64) {
+        match self.envelope {
+            Envelope::CosSquared => {
+                if phase.abs() < consts::PI * self.n_cycles() {
+                    let envelope = (phase / (2.0 * self.n_cycles())).cos().powi(4);
+                    (envelope, 2.0 * self.wavevector[0] * (phase / (2.0 * self.n_cycles())).tan() * envelope / self.n_cycles())
+                } else {
+                    (0.0, 0.0)
+                }
+            },
+
+            Envelope::Flattop => {
+                if phase.abs() > consts::PI * (self.n_cycles() + 1.0) {
+                    (0.0, 0.0)
+                } else if phase.abs() > consts::PI * (self.n_cycles() - 1.0) {
+                    let arg = 0.25 * (phase.abs() - (self.n_cycles() - 1.0) * consts::PI);
+                    (arg.cos().powi(4), self.wavevector[0] * phase.signum() * arg.sin() * arg.cos().powi(3))
+                } else {
+                    (1.0, 0.0)
+                }
+            },
+
+            Envelope::Gaussian => {
+                let tau = self.omega() * self.duration;
+                let envelope = (-4.0 * consts::LN_2 * phase.powi(2) / tau.powi(2)).exp();
+                (envelope, 8.0 * consts::LN_2 * self.wavevector[0] * phase * envelope / tau.powi(2))
+            }
+        }
     }
 
     pub fn a_sqd(&self, r: FourVector) -> f64 {
@@ -62,19 +118,7 @@ impl FocusedLaser {
 
         // Pulse envelope
         let phase = self.wavevector * r; // - r[3] * rho_sqd / (z_r * width_sqd);
-
-        #[cfg(feature = "cos2-envelope-in-3d")]
-        let envelope = if phase.abs() < consts::PI * self.duration {
-            (phase / (2.0 * self.duration)).cos().powi(4)
-        } else {
-            0.0
-        };
-
-        #[cfg(not(feature = "cos2-envelope-in-3d"))]
-        let envelope = {
-            let tau = self.omega() * self.duration;
-            (-4.0 * consts::LN_2 * phase.powi(2) / tau.powi(2)).exp()
-        };
+        let (envelope, _) = self.envelope_and_grad(phase);
 
         beam * envelope
     }
@@ -101,22 +145,7 @@ impl FocusedLaser {
 
         // Pulse envelope
         let phase = self.wavevector * r; // - r[3] * rho_sqd / (z_r * width_sqd);
-
-        #[cfg(feature = "cos2-envelope-in-3d")]
-        let (envelope, grad_envelope) = if phase.abs() < consts::PI * self.duration {
-            let envelope = (phase / (2.0 * self.duration)).cos().powi(4);
-            (envelope, 2.0 * self.wavevector[0] * (phase / (2.0 * self.duration)).tan() * envelope / self.duration)
-        } else {
-            (0.0, 0.0)
-        };
-
-        #[cfg(not(feature = "cos2-envelope-in-3d"))]
-        let (envelope, grad_envelope) = {
-            let tau = self.omega() * self.duration;
-            let envelope = (-4.0 * consts::LN_2 * phase.powi(2) / tau.powi(2)).exp();
-            (envelope, 8.0 * consts::LN_2 * self.wavevector[0] * phase * envelope / tau.powi(2))
-        };
-
+        let (envelope, grad_envelope) = self.envelope_and_grad(phase);
         let grad_envelope = [0.0, 0.0, grad_envelope];
 
         -FourVector::new(
@@ -137,33 +166,22 @@ impl FocusedLaser {
 }
 
 impl Field for FocusedLaser {
-    fn total_energy(&self) -> f64 {
-        // peak power of an LP Gaussian beam
-        let peak_field = ELECTRON_MASS * SPEED_OF_LIGHT * self.omega() * self.a0 / ELEMENTARY_CHARGE;
-        let peak_power = 0.25 * consts::PI * self.waist.powi(2) * peak_field.powi(2) / (SPEED_OF_LIGHT * VACUUM_PERMEABILITY);
-        // time = int f(t)^2 dt where f is the electric-field envelope
-        let time = self.duration * (consts::PI / 16.0f64.ln()).sqrt();
-        let norm = match self.pol {
-            Polarization::Linear => 1.0,
-            Polarization::Circular => 2.0,
-        };
-        norm * peak_power * time
-    }
-
     fn max_timestep(&self) -> Option<f64> {
-        Some(1.0 / self.omega())
+        let dt = match self.envelope {
+            Envelope::CosSquared | Envelope::Gaussian => 1.0 / self.omega(),
+            Envelope::Flattop => 0.2 / self.omega(),
+        };
+        Some(dt)
     }
 
-    #[cfg(feature = "cos2-envelope-in-3d")]
     fn contains(&self, r: FourVector) -> bool {
-        let phase: f64 = self.wavevector * r;
-        phase < consts::PI * self.duration
-    }
-
-    #[cfg(not(feature = "cos2-envelope-in-3d"))]
-    fn contains(&self, r: FourVector) -> bool {
-        let phase: f64 = self.wavevector * r;
-        phase < 3.0 * self.omega() * self.duration
+        let phase = self.wavevector * r;
+        let max_phase = match self.envelope {
+            Envelope::CosSquared => consts::PI * self.n_cycles(),
+            Envelope::Flattop => consts::PI * (self.n_cycles() + 1.0),
+            Envelope::Gaussian => 6.0 * consts::PI * self.n_cycles(), // 3.0 * self.omega() * self.duration
+        };
+        phase < max_phase
     }
 
     /// Advances particle position and momentum using a leapfrog method
@@ -247,6 +265,15 @@ impl Field for FocusedLaser {
             (prob, 0.0, None)
         }
     }
+
+    fn ideal_initial_z(&self) -> f64 {
+        let wavelength = 2.0 * consts::PI / self.wavevector[0];
+        match self.envelope {
+            Envelope::CosSquared => 0.5 * wavelength * self.n_cycles(),
+            Envelope::Flattop => 0.5 * wavelength * (self.n_cycles() + 1.0),
+            Envelope::Gaussian => 2.0 * wavelength * self.n_cycles(),
+        }
+    }
 }
 
 #[cfg(test)]
@@ -256,13 +283,15 @@ mod tests {
     #[test]
     fn on_axis() {
         let t_start = -20.0 * 0.8e-6 / (SPEED_OF_LIGHT);
-        let dt = 0.25 * 0.8e-6 / (SPEED_OF_LIGHT);
-        let laser = FocusedLaser::new(100.0, 0.8e-6, 4.0e-6, 30.0e-15, Polarization::Circular);
+        let n_cycles = SPEED_OF_LIGHT * 30.0e-15 / 0.8e-6;
+        let laser = FocusedLaser::new(100.0, 0.8e-6, 4.0e-6, n_cycles, Polarization::Circular)
+            .with_envelope(Envelope::Flattop);
+        let dt = laser.max_timestep().unwrap();
 
         let mut u = FourVector::new(0.0, 0.0, 0.0, -1000.0).unitize();
         let mut r = FourVector::new(0.0, 0.0, 0.0, 0.0) + u * SPEED_OF_LIGHT * t_start / u[0];
 
-        for _i in 0..(20*2*5) {
+        while laser.contains(r) {
             let new = laser.push(r, u, ELECTRON_CHARGE / ELECTRON_MASS, dt, EquationOfMotion::Lorentz);
             r = new.0;
             u = new.1;
@@ -282,7 +311,8 @@ mod tests {
         let w0 = 4.0e-6;
         let lambda = 0.8e-6;
         let gamma = 1000.0;
-        let laser = FocusedLaser::new(a0, lambda, w0, 30.0e-15, Polarization::Circular);
+        let n_cycles = SPEED_OF_LIGHT * 30.0e-15 / lambda;
+        let laser = FocusedLaser::new(a0, lambda, w0, n_cycles, Polarization::Circular);
 
         for k in 1..8 {
             let b = (k as f64) * 1.0e-6;