+++ src/field/mod.rs
@@ -36,6 +36,26 @@ impl Hdf5Type for Polarization {
     }
 }
 
+/// Temporal profile of the laser
+#[derive(Copy, Clone, Eq, PartialEq)]
+#[repr(u8)]
+pub enum Envelope {
+    CosSquared = 0,
+    Flattop = 1,
+    Gaussian = 2,
+}
+
+#[cfg(feature = "hdf5-output")]
+impl Hdf5Type for Envelope {
+    fn new() -> Datatype {
+        unsafe { Datatype::enumeration(&[
+            ("cos^2", Envelope::CosSquared as u8),
+            ("flattop", Envelope::Flattop as u8),
+            ("gaussian", Envelope::Gaussian as u8),
+        ])}
+    }
+}
+
 #[derive(Copy, Clone, Eq, PartialEq)]
 pub enum EquationOfMotion {
     Lorentz,
@@ -60,9 +80,6 @@ pub enum Laser {
 /// Represents the electromagnetic field in a spatiotemporal domain.
 #[enum_dispatch(Laser)]
 pub trait Field {
-    /// Returns the total electromagnetic energy in joules
-    fn total_energy(&self) -> f64;
-
     /// Returns the largest usuable value of the timestep `dt`
     /// used in the particle push, or `None` if there is no
     /// particular restriction
@@ -98,6 +115,10 @@ pub trait Field {
     /// otherwise be a rare event. The probability returned is *not*
     /// affected by this increase.
     fn pair_create<R: Rng>(&self, r: FourVector, ell: FourVector, pol: StokesVector, dt: f64, rng: &mut R, rate_increase: f64) -> (f64, f64, Option<(FourVector, FourVector, f64)>);
+
+    /// Returns `z0` such that an ultrarelatistic particle, initialized with `z = z0` at time `-z0/c`, is
+    /// sufficiently distant from the laser so as not to be affected by it.
+    fn ideal_initial_z(&self) -> f64;
 }
 
 #[cfg(test)]
@@ -107,8 +128,13 @@ mod tests {
 
     #[test]
     fn cp_deflection() {
-        let fast_laser = FastFocusedLaser::new(100.0, 0.8e-6, 4.0e-6, 30.0e-15, Polarization::Circular);
-        let laser = FocusedLaser::new(100.0, 0.8e-6, 4.0e-6, 30.0e-15, Polarization::Circular);
+        let n_cycles = 10.0;
+        let envelope = Envelope::Flattop;
+
+        let fast_laser = FastFocusedLaser::new(100.0, 0.8e-6, 4.0e-6, n_cycles, Polarization::Circular)
+            .with_envelope(envelope);
+        let laser = FocusedLaser::new(100.0, 0.8e-6, 4.0e-6, n_cycles, Polarization::Circular)
+            .with_envelope(envelope);
 
         let t_start = -20.0 * 0.8e-6 / (SPEED_OF_LIGHT);
         let x0 = 2.0e-6;
@@ -117,35 +143,45 @@ mod tests {
         let r = FourVector::new(0.0, x0, 0.0, 0.0) + u * SPEED_OF_LIGHT * t_start / u[0];
 
         // ponderomotive solver
-        let dt = 0.5 * 0.8e-6 / (SPEED_OF_LIGHT);
+        let dt = laser.max_timestep().unwrap();
         let mut pond = (r, u, dt);
-        for _i in 0..(20*2*2) {
+        while laser.contains(pond.0) {
             pond = laser.push(pond.0, pond.1, ELECTRON_CHARGE / ELECTRON_MASS, dt, EquationOfMotion::Lorentz);
         }
         let pond = pond.1;
 
         // Lorentz force solve
         // ponderomotive solver
-        let dt = 0.01 * 0.8e-6 / (SPEED_OF_LIGHT);
+        let dt = fast_laser.max_timestep().unwrap();
         let mut lorentz = (r, u, dt);
-        for _i in 0..(20*2*100) {
+        while fast_laser.contains(lorentz.0) {
             lorentz = fast_laser.push(lorentz.0, lorentz.1, ELECTRON_CHARGE / ELECTRON_MASS, dt, EquationOfMotion::Lorentz);
         }
         let lorentz = lorentz.1;
 
-        let theory = 2.0 * 3.63189;
+        let theory = 2.0 * match envelope {
+            Envelope::CosSquared => 1.13724,
+            Envelope::Flattop => 2.95684,
+            Envelope::Gaussian => 3.22816,
+        };
+
         let pond_angle = 1.0e3 * pond[1].atan2(-pond[3]);
         let lorentz_angle = 1.0e3 * lorentz[1].atan2(-lorentz[3]);
         let error = ((pond_angle - lorentz_angle) / lorentz_angle).abs();
 
-        println!("angle [PF] = {:.3e}, angle [LF] = {:.3e}, error = {:.3}%, predicted = {:.3e}", pond_angle, lorentz_angle, 100.0 * error, theory);
+        println!("angle [PF] = {:.3e} [{:.3e}], angle [LF] = {:.3e} [{:.3e}], error = {:.3}%, predicted = {:.3e}", pond_angle, 1.0e3 * pond[2].atan2(-pond[3]), lorentz_angle, 1.0e3 * lorentz[2].atan2(-lorentz[3]), 100.0 * error, theory);
         assert!(error < 1.0e-2);
     }
 
     #[test]
     fn lp_deflection() {
-        let fast_laser = FastFocusedLaser::new(100.0, 0.8e-6, 4.0e-6, 30.0e-15, Polarization::Linear);
-        let laser = FocusedLaser::new(100.0, 0.8e-6, 4.0e-6, 30.0e-15, Polarization::Linear);
+        let n_cycles = 10.0;
+        let envelope = Envelope::Gaussian;
+
+        let fast_laser = FastFocusedLaser::new(100.0, 0.8e-6, 4.0e-6, n_cycles, Polarization::Linear)
+            .with_envelope(envelope);
+        let laser = FocusedLaser::new(100.0, 0.8e-6, 4.0e-6, n_cycles, Polarization::Linear)
+            .with_envelope(envelope);
 
         let t_start = -20.0 * 0.8e-6 / (SPEED_OF_LIGHT);
         let y0 = 2.0e-6;
@@ -154,23 +190,28 @@ mod tests {
         let r = FourVector::new(0.0, 0.0, y0, 0.0) + u * SPEED_OF_LIGHT * t_start / u[0];
 
         // ponderomotive solver
-        let dt = 0.5 * 0.8e-6 / (SPEED_OF_LIGHT);
+        let dt = laser.max_timestep().unwrap();
         let mut pond = (r, u, dt);
-        for _i in 0..(20*2*2) {
+        while laser.contains(pond.0) {
             pond = laser.push(pond.0, pond.1, ELECTRON_CHARGE / ELECTRON_MASS, dt, EquationOfMotion::Lorentz);
         }
         let pond = pond.1;
 
         // Lorentz force solve
         // ponderomotive solver
-        let dt = 0.01 * 0.8e-6 / (SPEED_OF_LIGHT);
+        let dt = fast_laser.max_timestep().unwrap();
         let mut lorentz = (r, u, dt);
-        for _i in 0..(20*2*100) {
+        while fast_laser.contains(lorentz.0) {
             lorentz = fast_laser.push(lorentz.0, lorentz.1, ELECTRON_CHARGE / ELECTRON_MASS, dt, EquationOfMotion::Lorentz);
         }
         let lorentz = lorentz.1;
 
-        let theory = 3.63189;
+        let theory = match envelope {
+            Envelope::CosSquared => 1.13724,
+            Envelope::Flattop => 2.95684,
+            Envelope::Gaussian => 3.22816,
+        };
+
         let pond_angle = 1.0e3 * pond[2].atan2(-pond[3]);
         let lorentz_angle = 1.0e3 * lorentz[2].atan2(-lorentz[3]);
         let error = ((pond_angle - lorentz_angle) / lorentz_angle).abs();